Extract AreHardLinked out of utils:files
Bug: b/539666782
