+++ base/cvd/cuttlefish/common/libs/utils/BUILD.bazel
@@ -114,6 +114,7 @@ cf_cc_library(
         "//cuttlefish/common/libs/fs",
         "//cuttlefish/common/libs/utils:in_sandbox",
         "//cuttlefish/common/libs/utils:users",
+        "//cuttlefish/files:are_hard_linked",
         "//cuttlefish/files:copy",
         "//cuttlefish/files:file_device_id",
         "//cuttlefish/files:file_exists",
@@ -133,6 +134,7 @@ cf_cc_test(
     srcs = ["files_test.cpp"],
     deps = [
         "//cuttlefish/common/libs/utils:files",
+        "//cuttlefish/files:are_hard_linked",
         "//cuttlefish/files:file_exists",
         "//cuttlefish/result",
         "//cuttlefish/result:result_matchers",
