+++ base/cvd/cuttlefish/common/libs/utils/files_test.cpp
@@ -24,6 +24,7 @@
 #include "gmock/gmock.h"
 #include "gtest/gtest.h"
 
+#include "cuttlefish/files/are_hard_linked.h"
 #include "cuttlefish/files/file_exists.h"
 #include "cuttlefish/result/result.h"
 #include "cuttlefish/result/result_matchers.h"
